requirements.txt
Django==3.2.7
django-redis==5.0.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pytest-django==4.4.0
pytest~=6.2.5